Remove the explicit SyncClient constructor API
The hand-written client config (resources map, postChanges/postReplay POSTers, syncModel, auth/connectivity hooks) was boilerplate the framework can derive, so it is no longer possible to configure any of it by hand. new SyncClient(options) now derives everything from the configuration's registered models (static sync) and its sync.client block; options only carries configuration, legacyCursor, scopeStore, and syncModel overrides, and fromConfiguration delegates to the constructor. Specs drive the derived API through shared fakes in spec/sync/sync-client-fakes.js.
