+++ changelog.d/20260704171714-sync-client-from-configuration.md
@@ -1,1 +1,1 @@
-Derive the sync client from configuration and model declarations: models opt in with `static sync` (true or {track, syncType: "upsert", findRecord, localOnlyAttributes, ...}), and `SyncClient.fromConfiguration(...)` plus the lazy `syncClient()` accessor build the client from `configuration.getModelClasses()` — booleanAttributes from column types, localOnlyAttributes from the primary key/timestamps/sync bookkeeping columns, the registered `Sync` model as the pending-sync model, and framework-owned `/velocious/sync/changes|replay` POSTers over a new `sync.client = {transport, authenticationToken, isOnline, onError, batchSize}` configuration block. The default tracked payload now serializes Date attributes to ISO strings so per-model payload builders are unnecessary, and the `syncType: "upsert"` flag replaces function overrides for upsert-style resources.
+Derive the sync client entirely from configuration and model declarations: models opt in with `static sync` (true or {track, syncType: "upsert", findRecord, localOnlyAttributes, ...}), and `new SyncClient()` / `SyncClient.fromConfiguration(...)` / the lazy `syncClient()` accessor build the client from `configuration.getModelClasses()` — booleanAttributes from column types (driver-uniform, including MSSQL `bit`), localOnlyAttributes from the primary key/timestamps/sync bookkeeping columns, the registered `Sync` model as the pending-sync model, and framework-owned `${mountPath}/changes|replay` POSTers over a new `sync.client = {transport, authenticationToken, isOnline, onError, batchSize, mountPath}` configuration block. The old explicit constructor API (`resources` map, `postChanges`/`postReplay` POSTers, `syncModel`, auth/connectivity hooks) is removed — none of it can be hand-configured anymore. The default tracked payload now serializes Date attributes to ISO strings so per-model payload builders are unnecessary, and the `syncType: "upsert"` flag replaces function overrides for upsert-style resources.
